requirements.txt
asgiref==3.2.3
asn1crypto==1.3.0
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
Django==3.0.3
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
flake8==3.7.9
idna==2.9
image==1.5.28
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
whitenoise==5.0.1
gunicorn==20.0.4
django_heroku==0.3.1
django-crispy-forms==1.9.0